+++ app/Http/Controllers/Admin/SlideController.php
@@ -4,24 +4,43 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 use App\Models\Slide;
 
 class SlideController extends Controller
 {
-    public function index()
+    public function index(Request $request)
     {
+        if (!Auth::user()->can('listar-slides')) {
+            $request->session()->flash('mensagem',
+                ['msg'=>'Erro: Sem acesso à funcionalidade!', 'class'=>'red white-text']);
+            return redirect()->route('admin.home');
+        }
+
         $slides = Slide::orderBy('ordem')->get();
         return view('admin.slides.index', compact('slides'));
     }
 
-    public function cadastrar()
+    public function cadastrar(Request $request)
     {
+        if (!Auth::user()->can('cadastrar-slides')) {
+            $request->session()->flash('mensagem',
+                ['msg'=>'Erro: Sem acesso à funcionalidade!', 'class'=>'red white-text']);
+            return redirect()->route('admin.home');
+        }
+
         return view('admin.slides.cadastrar');
     }
 
     public function salvar(Request $request)
     {
+        if (!Auth::user()->can('cadastrar-slides')) {
+            $request->session()->flash('mensagem',
+                ['msg'=>'Erro: Sem acesso à funcionalidade!', 'class'=>'red white-text']);
+            return redirect()->route('admin.home');
+        }
+
         $dados = $request->all();
 
         if (Slide::count()) {
@@ -53,14 +72,26 @@ class SlideController extends Controller
         return redirect()->route('admin.slides');
     }
 
-    public function alterar($id)
+    public function alterar(Request $request, $id)
     {
+        if (!Auth::user()->can('atualizar-slides')) {
+            $request->session()->flash('mensagem',
+                ['msg'=>'Erro: Sem acesso à funcionalidade!', 'class'=>'red white-text']);
+            return redirect()->route('admin.home');
+        }
+
         $slide = Slide::find($id);
         return view('admin.slides.alterar', compact('slide'));
     }
 
     public function atualizar(Request $request, $id)
     {
+        if (!Auth::user()->can('atualizar-slides')) {
+            $request->session()->flash('mensagem',
+                ['msg'=>'Erro: Sem acesso à funcionalidade!', 'class'=>'red white-text']);
+            return redirect()->route('admin.home');
+        }
+
         $dados = $request->all();
 
         $slide = Slide::find($id);
@@ -104,6 +135,12 @@ class SlideController extends Controller
 
     public function remover(Request $request, $id)
     {
+        if (!Auth::user()->can('remover-slides')) {
+            $request->session()->flash('mensagem',
+                ['msg'=>'Erro: Sem acesso à funcionalidade!', 'class'=>'red white-text']);
+            return redirect()->route('admin.home');
+        }
+
         Slide::find($id)->delete();
 
         $request->session()->flash('mensagem',
